+++ src/volumito/__init__.py
@@ -13,7 +13,7 @@ from volumito.clients import (
     VolumioRESTAPIClient,
 )
 
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 __author__ = "Alberto Pettarin"
 __email__ = "alberto@albertopettarin.it"