+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.9] - 2026-07-20
+
+### Added
+
+- Short options for frequently used options: `-H` (`--host`), `-M` (`--mpd-port`),
+  `-P` (`--rest-api-port`), `-p` (`--position`), `-F` (`--format`), `-L` (`--fields`),
+  and `-R` (`--raw`)
+
+
 ## [0.0.8] - 2026-07-19
 
 ### Changed